+++ gems/aws-sdk-s3/spec/plugins/arn_spec.rb
@@ -3,6 +3,9 @@ require_relative '../spec_helper'
 module Aws
   module S3
     describe Client do
+      # squash legacy fips shim transform warnings
+      before { allow(Aws::Plugins::RegionalEndpoint).to receive(:warn) }
+
       describe 's3_use_arn_region' do
         it 'is configured to use the arn region by default' do
           client = Aws::S3::Client.new(
@@ -339,15 +342,43 @@ module Aws
           end.to raise_error(Aws::Errors::InvalidARNRegionError)
         end
 
-        it 'raises when the fips client region does not match the arn region' do
+        it 'uses the arn region with the fips client region' do
           client = Aws::S3::Client.new(
             stub_responses: true,
             region: 'fips-us-gov-east-1'
           )
           arn = 'arn:aws-us-gov:s3:us-gov-west-1:123456789012:accesspoint:myendpoint'
-          expect do
-            client.get_object(bucket: arn, key: 'obj')
-          end.to raise_error(Aws::Errors::InvalidARNRegionError)
+          resp = client.get_object(bucket: arn, key: 'obj')
+          host = 'myendpoint-123456789012.s3-accesspoint-fips.us-gov-west-1.amazonaws.com'
+          expect(resp.context.http_request.endpoint.host).to eq(host)
+          expect(resp.context.http_request.endpoint.path).to eq('/obj')
+        end
+
+        it 's3_use_arn_region false; uses the fips configuration flag' do
+          client = Aws::S3::Client.new(
+            stub_responses: true,
+            region: 'us-gov-east-1',
+            use_fips_endpoint: true,
+            s3_use_arn_region: false
+          )
+          arn = 'arn:aws-us-gov:s3:us-gov-east-1:123456789012:accesspoint:myendpoint'
+          resp = client.get_object(bucket: arn, key: 'obj')
+          host = 'myendpoint-123456789012.s3-accesspoint-fips.us-gov-east-1.amazonaws.com'
+          expect(resp.context.http_request.endpoint.host).to eq(host)
+          expect(resp.context.http_request.endpoint.path).to eq('/obj')
+        end
+
+        it 'uses the fips configuration flag' do
+          client = Aws::S3::Client.new(
+            stub_responses: true,
+            region: 'us-gov-east-1',
+            use_fips_endpoint: true
+          )
+          arn = 'arn:aws-us-gov:s3:us-gov-west-1:123456789012:accesspoint:myendpoint'
+          resp = client.get_object(bucket: arn, key: 'obj')
+          host = 'myendpoint-123456789012.s3-accesspoint-fips.us-gov-west-1.amazonaws.com'
+          expect(resp.context.http_request.endpoint.host).to eq(host)
+          expect(resp.context.http_request.endpoint.path).to eq('/obj')
         end
 
         it 'uses the arn region for a fips client with dualstack' do
@@ -917,21 +948,49 @@ module Aws
             region: 'fips-us-gov-east-1',
             s3_use_arn_region: false
           )
-          arn = 'arn:aws-us-gov:s3:us-gov-west-1:123456789012:accesspoint:myendpoint'
+          arn = 'arn:aws-us-gov:s3-object-lambda:us-gov-west-1:123456789012:accesspoint/mybanner'
           expect do
             client.get_object(bucket: arn, key: 'obj')
           end.to raise_error(Aws::Errors::InvalidARNRegionError)
         end
 
-        it 'raises when the fips client region does not match the arn region' do
+        it 'uses the arn region with the fips client region' do
           client = Aws::S3::Client.new(
             stub_responses: true,
             region: 'fips-us-gov-east-1'
           )
-          arn = 'arn:aws-us-gov:s3:us-gov-west-1:123456789012:accesspoint:myendpoint'
-          expect do
-            client.get_object(bucket: arn, key: 'obj')
-          end.to raise_error(Aws::Errors::InvalidARNRegionError)
+          arn = 'arn:aws-us-gov:s3-object-lambda:us-gov-west-1:123456789012:accesspoint/mybanner'
+          resp = client.get_object(bucket: arn, key: 'obj')
+          host = 'mybanner-123456789012.s3-object-lambda-fips.us-gov-west-1.amazonaws.com'
+          expect(resp.context.http_request.endpoint.host).to eq(host)
+          expect(resp.context.http_request.endpoint.path).to eq('/obj')
+        end
+
+        it 's3_use_arn_region false; uses the fips configuration flag' do
+          client = Aws::S3::Client.new(
+            stub_responses: true,
+            region: 'us-gov-east-1',
+            use_fips_endpoint: true,
+            s3_use_arn_region: false
+          )
+          arn = 'arn:aws-us-gov:s3-object-lambda:us-gov-east-1:123456789012:accesspoint/mybanner'
+          resp = client.get_object(bucket: arn, key: 'obj')
+          host = 'mybanner-123456789012.s3-object-lambda-fips.us-gov-east-1.amazonaws.com'
+          expect(resp.context.http_request.endpoint.host).to eq(host)
+          expect(resp.context.http_request.endpoint.path).to eq('/obj')
+        end
+
+        it 'uses the fips configuration flag' do
+          client = Aws::S3::Client.new(
+            stub_responses: true,
+            region: 'us-gov-east-1',
+            use_fips_endpoint: true
+          )
+          arn = 'arn:aws-us-gov:s3-object-lambda:us-gov-west-1:123456789012:accesspoint/mybanner'
+          resp = client.get_object(bucket: arn, key: 'obj')
+          host = 'mybanner-123456789012.s3-object-lambda-fips.us-gov-west-1.amazonaws.com'
+          expect(resp.context.http_request.endpoint.host).to eq(host)
+          expect(resp.context.http_request.endpoint.path).to eq('/obj')
         end
 
         it 'raises with :use_accelerate_endpoint' do