+++ gems/aws-sdk-s3control/lib/aws-sdk-s3control/arn/outpost_bucket_arn.rb
@@ -28,13 +28,18 @@ module Aws
 
       def validate_arn!
         unless @service == 's3-outposts'
-          raise ArgumentError, 'Must provide a valid S3 outposts bucket ARN.'
+          raise ArgumentError, 'Must provide a valid S3 Outpost Bucket ARN.'
         end
 
         if @region.empty? || @account_id.empty?
           raise ArgumentError,
-                'S3 accesspoint ARNs must contain both a region '\
-                'and an account id.'
+                'S3 Outpost Bucket ARNs must contain both a region '\
+                'and an Account ID.'
+        end
+
+        if @region.include?('-fips') || @region.include?('fips-')
+          raise ArgumentError,
+                'S3 Outpost Bucket ARNs cannot contain a FIPS region.'
         end
 
         if @type != 'outpost' && @subtype != 'bucket'
@@ -42,16 +47,16 @@ module Aws
         end
 
         if @outpost_id.nil? || @outpost_id.empty?
-          raise ArgumentError, 'Missing ARN outpost id.'
+          raise ArgumentError, 'Missing ARN Outpost ID.'
         end
 
         if @bucket_name.nil? || @bucket_name.empty?
-          raise ArgumentError, 'Missing ARN accesspoint name.'
+          raise ArgumentError, 'Missing ARN bucket name.'
         end
 
         if @extra
           raise ArgumentError,
-                'ARN outpost bucket must be a single value.'
+                'ARN Outpost bucket must be a single value.'
         end
 
         unless Seahorse::Util.host_label?(@outpost_id)
@@ -64,11 +69,11 @@ module Aws
       end
 
       # Outpost Bucket ARNs currently do not support dualstack
-      def host_url(region, _dualstack = false, custom_endpoint = nil)
+      def host_url(region, fips = false, _dualstack = false, custom_endpoint = nil)
         if custom_endpoint
           custom_endpoint
         else
-          "s3-outposts.#{region}.amazonaws.com"
+          "s3-outposts#{'-fips' if fips}.#{region}.amazonaws.com"
         end
       end
     end